virtio: queue_reset: pci: support VIRTIO_F_RING_RESET
This patch implements virtio pci support for QUEUE RESET.

Performing reset on a queue is divided into three steps:

1. reset_vq: notify the hardware queue to reset
2. del_reset_vq: delete the reset queue
3. enable_reset_vq: re-enable the reset queue

Between steps 1 and 2, generally call virtqueue_detach_unused_buf() to
recycle the buffer.

When deleting a vq, vp_del_vq() will be called to release all the memory
of the vq. But this does not affect the process of deleting vqs, because
that is based on the queue to release all the vqs. During this process,
the vq has been removed from the queue.

When deleting vq, info and vq will be released, and I save msix_vec in
vp_dev->vqs[queue_index]. When re-enable, the current msix_vec can be
reused. And based on intx_enabled to determine which method to use to
enable this queue.

Signed-off-by: Xuan Zhuo <xuanzhuo@linux.alibaba.com>